Improve formatting of JSON diff
When adding or removing an animation, don't show info for all of the other
animations that were shifted up or down in the animation list.

This does mean that we won't show diff contents if animations were simply
re-ordered.  However, this seems like a reasonable trade-off for now.
Adding or removing animations seems more common than re-ordering them.